Implemnt authorization using policy

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,12 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+
+    public function __constructor(Type $var = null)
+    {
+        $this->middleware('auth',['except'=> 'index','show']);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -27,6 +33,7 @@ class QuestionsController extends Controller
      */
     public function create()
     {
+
         $question = new Question();
         return view('questions.create',compact('question'));
     }
@@ -66,6 +73,7 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        $this->authorize("update",$question);
         return view('questions.edit',compact('question'));
     }
 
@@ -78,6 +86,7 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update",$question);
     $question->update($request->only('title','body'));
 
         return redirect()->route('questions.index')->with('success', "Your question has been updated.");
@@ -91,6 +100,7 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize("delete",$question);
         $question->delete();
         return redirect()->route('questions.index')->with('success', "Your question has been deleted.");
